Extract vtable.rs
